increased num of elements for combine

+++ compiler-optimizations/combine.h
@@ -1,11 +1,11 @@
 #ifndef COMBINE
 #define COMBINE
 
-void combine3(int* data, int data_len, int* dest);
-void combine4(int* data, int data_len, int* dest);
-void combine5(int* data, int data_len, int* dest);
-void combine6(int* data, int data_len, int* dest);
-void combine7(int* data, int data_len, int* dest);
+void combine3(int* data, long data_len, long* dest);
+void combine4(int* data, long data_len, long* dest);
+void combine5(int* data, long data_len, long* dest);
+void combine6(int* data, long data_len, long* dest);
+void combine7(int* data, long data_len, long* dest);
 
 
 #endif
